+++ src/routes/characters.py
@@ -2,7 +2,7 @@
 
 from flask import Response, jsonify, make_response, request
 
-import database as db
+import database_old as db
 
 from .main import bp
