+++ src/routes/messages.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta, timezone
 
 from flask import Response, g, jsonify, make_response, request
 
-import database as db
+import database_old as db
 
 from .main import bp
