+++ src/routes/users.py
@@ -3,7 +3,7 @@
 from flask import Response, jsonify, make_response, request
 
 import auth
-import database as db
+import database_old as db
 
 from .main import bp
