+++ tests/test_auth/test_passwords.py
@@ -7,9 +7,9 @@ This file contains the tests for the auth/passwords.py file.
 import pytest
 
 import auth
-import database as db
-from tests.test_database.test_main import db_init
-from tests.test_database.test_users import user_1
+import database_old as db
+from tests.test_database_old.test_main import db_init
+from tests.test_database_old.test_users import user_1
 
 
 def test_insert_user(db_init: str) -> None:
