+++ tests/test_database_old/test_messages.py
@@ -10,11 +10,11 @@ from typing import Generator
 
 import pytest
 
-import database as db
-from tests.test_database.test_characters import char_1, char_2
-from tests.test_database.test_main import db_init
-from tests.test_database.test_threads import thread_1, thread_2
-from tests.test_database.test_users import user_1, user_2
+import database_old as db
+from tests.test_database_old.test_characters import char_1, char_2
+from tests.test_database_old.test_main import db_init
+from tests.test_database_old.test_threads import thread_1, thread_2
+from tests.test_database_old.test_users import user_1, user_2
 
 
 @pytest.fixture
