+++ tests/test_routes/test_events.py
@@ -7,11 +7,11 @@ This file contains the tests for the routes/events.py file.
 
 from flask.testing import FlaskClient
 
-import database as db
+import database_old as db
 from tests.test_app import app, client
-from tests.test_database.test_characters import char_1
-from tests.test_database.test_events import event_1, event_2
-from tests.test_database.test_main import db_init
+from tests.test_database_old.test_characters import char_1
+from tests.test_database_old.test_events import event_1, event_2
+from tests.test_database_old.test_main import db_init
 
 
 def test_get_events(client: FlaskClient, event_1: db.Event, event_2: db.Event) -> None:
