+++ tests/test_routes/test_posts.py
@@ -6,11 +6,11 @@ This file contains the tests for the routes/posts.py file.
 
 from flask.testing import FlaskClient
 
-import database as db
+import database_old as db
 from tests.test_app import app, client
-from tests.test_database.test_characters import char_1
-from tests.test_database.test_main import db_init
-from tests.test_database.test_posts import post_1, post_2
+from tests.test_database_old.test_characters import char_1
+from tests.test_database_old.test_main import db_init
+from tests.test_database_old.test_posts import post_1, post_2
 
 
 def test_get_posts(client: FlaskClient, post_1: db.Post, post_2: db.Post) -> None:
